fix: Remove unnecessary (use client) directive from use-cases/page.tsx
